+++ Makefile
@@ -1,2 +1,2 @@
 all:
-	go build -o bin/event-to-sentry *.go && ./bin/event-to-sentry
+	go build -o bin/main *.go && ./bin/main -i
